cmake debug option (-g3); default linker no xram
Change teh default options for debug build from -g to -g3. This options
make the compiler emit more informations, e.g. about macros
Change linker script to not use the xram by default. Remember that this
change is because ram is supposed to be unreliable in the current
branch, and for testing must remain unused by the system